[Refactoring] split location context
